+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.6
 
+- samples/cairo/clip_image.rb
+- samples/cairo/curve_to.rb
 - Support declarative `fill`/`paint` of `image, x, y` in cairo graphics
 - Support declarative transforms on cairo graphics shapes: `translate`, `scale`, `rotate`
 - Support applying transform on `drawing_area` `paint`